fixed models

+++ src/transformers/models/layoutlmv3/modeling_layoutlmv3.py
@@ -910,8 +910,8 @@ class LayoutLMv3Model(LayoutLMv3PreTrainedModel):
         patch_height = patch_width = None
         if pixel_values is not None:
             patch_height, patch_width = (
-                int(pixel_values.shape[2] / self.config.patch_size),
-                int(pixel_values.shape[3] / self.config.patch_size),
+                (pixel_values.shape[2] / self.config.patch_size).to(torch.int64),
+                (pixel_values.shape[3] / self.config.patch_size).to(torch.int64),
             )
             visual_embeddings = self.forward_image(pixel_values)
             visual_attention_mask = torch.ones(
